+++ fau-budgeting/BudgetingDb.dbml
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?><Database Name="fau-budgeting" Class="BudgetingDbDataContext" xmlns="http://schemas.microsoft.com/linqtosql/dbml/2007">
-  <Connection Mode="WebSettings" ConnectionString="Data Source=DESKTOP-1K0C0QJ;Initial Catalog=fau-budgeting;Integrated Security=True" SettingsObjectName="System.Configuration.ConfigurationManager.ConnectionStrings" SettingsPropertyName="fau_budgetingConnectionString" Provider="System.Data.SqlClient" />
+  <Connection Mode="WebSettings" ConnectionString="Data Source=localhost;Initial Catalog=fau-budgeting;Integrated Security=True" SettingsObjectName="System.Configuration.ConfigurationManager.ConnectionStrings" SettingsPropertyName="fau_budgetingConnectionString" Provider="System.Data.SqlClient" />
   <Table Name="dbo.BudgetRequests" Member="BudgetRequests">
     <Type Name="BudgetRequest">
       <Column Name="Id" Type="System.Int32" DbType="Int NOT NULL IDENTITY" IsPrimaryKey="true" IsDbGenerated="true" CanBeNull="false" />
@@ -8,6 +8,7 @@
       <Column Name="RequestType" Type="System.String" DbType="VarChar(50) NOT NULL" CanBeNull="false" />
       <Column Name="OrganizationId" Type="System.Int32" DbType="Int NOT NULL" CanBeNull="false" />
       <Column Name="RequestData" Type="System.String" DbType="Text NOT NULL" CanBeNull="false" UpdateCheck="Never" />
+      <Column Member="Comments" Type="System.String" CanBeNull="true" />
       <Association Name="Organization_BudgetRequest" Member="Organization" ThisKey="OrganizationId" OtherKey="Id" Type="Organization" IsForeignKey="true" />
     </Type>
   </Table>
