+++ fau-budgeting/content/js/budget-info.js
@@ -1,0 +1,27 @@
+﻿function displayBudgetInfo(id) {
+    $.ajax({
+        url: "/api/budget-request-info",
+        data: {
+            id: id
+        },
+        success: function (response) {
+            var formData = JSON.parse(response);
+
+            $.each(formData, function (key, value) {
+                $("input[name='" + key + "']").attr("value", value);
+            });
+        }
+    });   
+}
+
+function disableEditing() {
+    $("input").each(function () {
+        $(this).prop("readonly", true);
+    });
+}
+
+function enableEditing() {
+    $("input").each(function () {
+        $(this).prop("readonly", false);
+    });
+}
