+++ fau-budgeting/views/organization/organization.sshtml
@@ -22,9 +22,9 @@
                     <div class="panel-body">
                         <div class="list-group">
                             @Each
-                                <a href="/budget-request-info?id=@Current.Id" class="list-group-item">
+                                <a href="/budget-request-info/organization?id=@Current.Id" class="list-group-item">
                                     <p>
-                                        <span class="request">@Current.RequestType</span> <br> <span class="@Current.Html.Class">@Current.Status<br></span>
+                                        <span class="em">@Current.RequestType</span> <br> <span class="@Current.Html.Class">@Current.Status<br></span>
                                         <span class="pull-right text-muted small" class="strong em">@Current.Date</span>
                                     </p>
                                 </a>
@@ -37,8 +37,4 @@
         </div>
 
     </div>
-@EndSection
-
-@Section['Styles']
-    <link href="/dist/css/sb-admin-2.css" rel="stylesheet">
 @EndSection
